SDP-488 (Chore): refactor VerifyReceiverRegistrationHandler (#32)
### What

Refactor VerifyReceiverRegistrationHandler.

Also, added some godoc and rename a method.

### Why

It was doing tens of calls in sequence that had to be maintained in the correct sequence. I split them into 4 private methods for make it easier to read and to test.

My task involves storing the AP transactionID in the database, and testing that part was troublesome in the monolithic method.