data/httphandler: add sms_resend_interval to organizations table (#65)
What
This PR adds a new column on the organizations table - sms_resend_interval. Also, adds this field in the PATCH /organizations endpoint.

Why
This new column will control the interval of resending the Invitation SMS to the receivers when they don't get registered after receiving the first invitation. The value of this column can be null or an integer greater than 0 (zero). Null means that the organization doesn't want to automatically resend the invitation SMS. To set this field as null a request with the number 0 (zero) should be sent.