[SDP-913] stellar-multitenant: add-tenants cmd (#76)
What
This PR adds a new CLI command in the stellar-multitenant package. This CLI is responsible for registering new tenants on the main database schema. The tenant name is unique.

```sh
mtn add-tenants myorg
```

Why
We should be able to register new tenants.

Known limitations
This process only registers the new tenant in the main database schema but we are missing creating the database schema and configuring the basics for it.