[SDP-913] stellar-multitenant: add config-tenant command (#77)
### What

This PR adds the `config-tenant` CLI command in the `stellar-multitenant` package. This command sets the basic environment variables for the tenants individually. Each flag of the command is optional, but at least one should be provided when running it. When a flag is not provided the corresponding value in the database will not be changed. To change a value it should be explicitly passed when running the CLI. 

The `tenant-id` is required to run the CLI.


```sh
mtn config-tenant --tenant-id abc --enable-mfa "false" --distribution-public-key "GBDYRFO3ZJ74UQ6V36D2PHXRXBYW52HKUYRYCWSFIGWQSPJMN4ARBGIF"
```

### Why

We should be able to configure the tenants' environment variables.

### Known limitations

We are still not able to set up `secrets`.
Missing tests for the CLI.