[SDP-953] cmd: change the migrate CLI to be tenant aware (#89)
### What

This PR modifies the SDP's `db` CLI command and `setup-for-network` CLI command to be tenant-aware. It means being able to run these commands for all or a specific tenant.

### Why

We should be able to apply the migrations to the tenants and also setup it for the given network.

### Known limitations

Fix the e2e test and SDP anchor platform compatibility.