[SDP-962] Change SEP-24 Flow to display different verifications based on Disbursement's verification type (#116)
Modify the SEP-24 flow to perform verification for an entered phone number based on the latest verification type.

The current SEP-24 flow is hardcoded to only accept date of birth but we will have disbursement files that will include pin and national id, and the front-end will need to change to be able to parse those values.

Release 1.0.1 to develop (#129)

[Release 1.0.1](https://github.com/stellar/stellar-disbursement-platform-backend/issues/127) to develop

To sync the `main` branch hotfixes:
- #125
- #126

changes:

fix tests

migrate