[SDP-922] events/serve: kafka setup (#119)
What
This PR adds the initial setup for the message broker's communication. This implementation uses Apache Kafka as the broker. The client Kafka used is the segmention/kafka-go package because it uses pure go and doesn't have any infra dependency to use it. The docker image used to run the Kafka locally is the bitnami/kafka.

Why
We are going to use Kafka to run services asynchronously per tenant.