[Hot-fix] TSS amount precision (#176)
### What

Fix TSS's database amount precision to be compliant with the Stellar network amount precision.

### Why

The database amount was configured to NUMERIC(10,7), which allows the maximum value of 999.9999999. This is not compliant with the Stellar network amounts though, which according with the [docs](https://developers.stellar.org/docs/fundamentals-and-concepts/stellar-data-structures/assets#amount-precision) supports numbers from 0.0000001 to 922,337,203,685.4775807.