[SDP-1040] httphandler/events/utils: improve kafka error handling when downtime (#171)
What
This PR adds an exponential backoff when there are Kafka errors before trying again. Also, it ensures we roll back to the previous states when fail to write new messages on Kafka.

Why
We need to ensure the Kafka reconnection when downtime and the data integrity when errors occurs.