[SDP-1112] inject dbConnectionPool as a dependency using the `dependencyinjection` package (#207)
### What

Update how the database connection pool is provided to the multiple services and commands by using dependency injection rather than passing the DSN around. This included:
* adding dependency injection (`di`) methods to start the Admin, TSS and MTN connection pools
* Metrics are supported in the `di` methods for Admin and TSS
* Use the `di` constructors in cmd/serve.go, cmd/channel_accounts.go, cmd/transaction_submitter.go, mtn/cli/add_tenants.go, mtn/cli/config_tenant.go

### Why

Now that we have multiple database connection pools, it gets easier to mistakenly instantiate them differently throughout the code. By centralizing where the instances are created, things get easier to maintain and to keep consistent.

Also, this guarantees that we share the same connection pool among different services when possible/desired

For instance, we were instantiating a new connection pool for every kafka handler (+3), and for the scheduled jobs (+1), and for two of the servers launched under the serve command (+2), which was creating 6 connections to the same database without need. This wastes resources and can overload the maximum number of connections as things scale further.

### Further Changes

* Implement DeleteInstanceBy(Value|Key) in the dependencyinjection package, to take care of cleaning dependencies from the di container and calling the close command on databases if needed.
* Fix the tss monitorService, so I could use it
* Remove flag `"enable-multitenant-db"`

### Known limitations

Further changes that were out of context were marked with TODOs and their tickets (SDP-874, SDP-968, SDP-1074) were mentioned here. I also added a note to those tickets' descriptions about these TODOs.