[SDP-1100] middleware/stellar-multitenant/serve/cmd: add default tenant (#249)
What
This PR adds the single tenant mode feature. A new environment variable was added SINGLE_TENANT_MODE and a new column was added to the public.tenants table is_default.

It's now possible to set the env var SINGLE_TENANT_MODE to true and then configure a tenant to be the default tenant by setting the is_default to true as well.

After this, every request will be resolved to the default tenant regardless of the Sdp-Tenant-Name request header or the Hostname prefix. This is useful when configuring the SDP to run with only one tenant.

Why
Make the setup for a single-tenant configuration easier.