[SDP-1209] Validate TENANT_XLM_BOOTSTRAP_AMOUNT on server startup (#291)
### What

Validate TENANT_XLM_BOOTSTRAP_AMOUNT on server startup. This validation was added to the `provisioning.NewManager` method.

Also, refactored the test `Test_Manager_ProvisionNewTenant`

### Why

If a server is misconfigured, it needs to be surfaced on server launch, rather than leaving the error to be triggered and debugged by a host at a later time.

Address https://stellarorg.atlassian.net/browse/SDP-1209.