[Fix][SDP-1274] Deactivate distribution acc when Circle API  returns 401 (#359)
When the Circle API actually responds with a 401, the error handler doesn't actually modify the distribution account status on the tenant because the condition for the tenant status is incorrect (we don't actually set the status to TENANT_ACTIVATED via the FE) so no tenant is selected. This fix changes the query so that it will attempt to select the tenant regardless of tenant status.