SDP-1569: Change `{CIRCLE_API}/ping` method to validate only the response status code (#580)
### What

Change `{CIRCLE_API}/ping` method to validate only the response status code

### Why

It was validating the response body as well, but Circle inadvertently changed it and the API is not compliant with the documentation anymore. Since this is just a /ping endpoint, we don't need to enforce any specific body content.