SDP-1763 add sep24 info endpoint (#773)
### What

Add the **GET** `/sep24/info` endpoint to return the info about
available assets
**Response:**
```
{ 
 "deposit": {
  "native": { "enabled": true, "min_amount": 1, "max_amount": 10000 }, 
  "USDC": { "enabled": true, "min_amount": 1, "max_amount": 10000 } 
}, 
 "withdraw": {}, 
 "fee": { "enabled": false }, 
 "features": { "account_creation": false, "claimable_balances": false } 
}
```

### Why

As part of the SEP24 migration to the SDP, we need to comply the SEP
requirement

### Known limitations

the values of the `min_amount` and `max_amount` are hardcoded as well as
the `fee`, `features`, `withdraw`

### Checklist

- [ ] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [ ] PR has a focused scope and doesn't mix features with refactoring
- [ ] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production