Fix: Only check transaction hash for successful wallet creation transactions (#805)
### What

This updates the TSS -> embedded wallet table sync job so that the
Stellar transaction hash is only checked if the transaction was
successfully submitted. This is necessary for wallet creation
transactions, as TSS can mark the transaction as an error during
simulation.

### Why

This fixes a bug where TSS records fail to sync to the embedded wallets
table.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production