SDP-1813 - [Kaizen] Refactor packages to reduce circular dependency issues (#816)
> [!NOTE]
> ### PR  [1 / 2] 
> I'm splitting this work to 2 PRs to make it easy to review. This PR
refactors the Tenant and TenantStatus struct. Next PR will move some of
the context related methods to another low-level package that will
reduce the dependency on the `tenant` package across the codebase.



### What
- Move `Tenant` and `TenantStatus` structs to `schema` package.

### Why
These structs are shared across the codebase. Moving them to `schema`
reduces dependencies on the `tenant` package which reduces the circular
dependency issue.

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [x] `CHANGELOG.md` is updated (if applicable)
- [x] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [x] Preview deployment works as expected
- [x] Ready for production