SDP 1812 add api keys caching (#815)
### What

Add api keys caching mechanism to reduce the number of calls to the DB,
this solution adds caching layer to reduce db load and at the same time
maintains the last update timestamp

### Why

In the current impl we have to hits per request, check the key exists
and update it's last used timestamp

### Known limitations

[TODO or N/A]

### Checklist

- [ ] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [ ] PR has a focused scope and doesn't mix features with refactoring
- [ ] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production