SDP-1817 Enable direct payments to SEP-24 wallets (#860)
### What

Added additional checks to the direct payment service, in case if the
unregistered wallet selected, the event with invitation will be emitted
and wallet with READY status created awaiting for the registration to be
finished.

UI:
https://github.com/stellar/stellar-disbursement-platform-frontend/pull/343

### Why

Currently, when creating a Direct Payment, only wallets already
associated with the receiver are displayed. This creates an issue for
receivers created through Direct Receiver Onboarding, as they don't have
any associated wallets.

### Known limitations

[TODO or N/A]

### Checklist

- [ ] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [ ] PR has a focused scope and doesn't mix features with refactoring
- [ ] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production