SDP-1871: Add Twilio WhatsAPP template documentation (#877)
### What
- Add documentation to explain what templates need to be created on
Twilio side to support the SDP.
- Removed extra templates that are currently not supported. I originally
put them in there to future proof the message sending in case we decide
to support Twilio Whatsapp for User communication, but that adds a lot
of config overhead that is unnecessary. Decided to remove it and can put
it back later.

### Why
- Help with and simplify setting up Twilio WhatsAPP for the SDP. 


### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production