SDP-1862: Implement passkey registration + authentication (#890)
### What

This implements the passkey registration + authentication flows. The
registration flow is called to get the credential ID, and public key
needed to call the embedded wallet API. The authentication flow is
called to get embedded wallet associated with a passkey and to generate
the JWT the wallet will use to access authenticated APIs.

The HTTP handler + dependency injection will be added in a follow-up PR.

### Why

Implements the passkey related APIs for the wallet frontend.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production