SDP-1615: Implement sponsored transaction TSS worker (#878)
### What

This implements the sponsored transaction TSS worker which picks up
signed invoke host function operations from an embedded wallet and
submits them to the network. This worker can also be refactored to be
generic enough so that the payment and wallet creation workers use this.

### Why

Sponsored transactions will be submitted through the TSS.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production