SDP-1615: Add sponsored transactions service + handler (#886)
### What

This adds the sponsored transaction endpoints for embedded wallets.
Authentication using JWT will be added to the endpoints at a later time
so account validation is left as a TODO.

### Why

We want to sponsor transactions for embedded wallets.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production