SDP-1617: Add refresh token endpoint for embedded wallets (#907)
### What

This adds a refresh endpoint that embedded wallets will call to extend
the expiration of their JWTs. It also updates the
`/passkey/registration/finish` endpoint to return a JWT so that the
wallet can immediately be authenticated without requiring the user to
authenticate again. This JWT will initially not contain a contract
address in the claims, but refreshing the token will populate it once
the wallet is created.

### Why

The 15-minute expiration of the embedded wallet JWT is too short without
a refresh mechanism.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production