SDP-1615: Add sponsored transaction sync job (#912)
### What

This adds the sponsored transaction from the TSS transaction scheduled
job. The implementation and unit tests are more or less a copy of the
wallet creation scheduled job.

### Why

We need to sync the status of the sponsored transaction with the TSS
transaction.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production