SDP-1883: Enable short linking by default (#916)
### What

set `is_link_shortener_enabled` default to true

### Why

Due to user feedback, the long deep links are expensive to send via SMS.
Currently “Enable Short Linking” is set to OFF by default and it should
be set to ‘Turn on’.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [ ] Tests are included (if applicable)
- [x] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production