SDP-1869: Remove events from TSS (#919)
### What

This removes event publishing from TSS.

### Why

It's difficult to resolve merge conflicts from merging the develop
branch since this feature branch builds on top of the event
infrastructure. It's safer to remove it first before merging.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production