SDP-1620: Fix sponsored transaction submission (#927)
### What

This fixes the sponsored transaction submission by allowing
`InvokeHostFunctionOp` to be submitted instead of an `Operation`, and
fixes the TSS record insert by using the correct connection pool.

### Why

N/A

### Known limitations

Using a separate connection pool for sponsored transactions and the TSS
transaction means that the inserts are not part of the same transaction.
I left a TODO here to implement the outbox/scheduler pattern similar to
payments.

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production