Separate Nginx proxy from frontend container (#937)
### What

This separates the Nginx proxy from the frontend container so that the
FE container can be stopped when doing frontend development with HTTPS
enabled.

Now you just need to:
1. (BE) `./dev/main.sh --https`
2. (FE) yarn start
3. Kill FE container

### Why

Make frontend development with HTTPS easier.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [ ] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production