SDP-1863: Implement SEP-45 challenge validation (#946)
### What

This implements the SEP-45 challenge validation flow
https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0045.md#token

### Why

The Anchor Platform has been removed, so we need to reimplement SEP-45.

### Known limitations

The HTTP handler, dependency injection, and JWT generation will be
implemented separately to keep the size of this PR reasonable.

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production