Add HTTPS mode to setup wizard (#957)
### What

This adds an HTTPS mode to the setup wizard. The protocol is set at the
.env level, meaning a profile created with HTTPS cannot be used over
HTTP later, and vice versa.

### Why

HTTPS is required to work with passkeys in a multitenant setup. It's
been implemented in the `feature/c-accounts` branch in the `dev/main.sh`
script, but needs to be ported to the setup wizard.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [ ] Tests are included (if applicable)
- [x] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production