Fix: Remove SEP-45 client domain attribution and fix server signing (#981)
### What

This removes the SEP-45 client attribution requirement since embedded
wallets currently do not have a concept of a client domain. It also
fixes a bug where the server auth entry is not present in the challenge
auth entries.

### Why

Blocking SEP-45 client-side implementation.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production