SDP-1923: Implement supported assets endpoint (#984)
### What

This implements the supported assets endpoint, which the embedded wallet
will call to fetch balances.

### Why

Trustlines do not exist for contract accounts, so we need to tell the
frontend which assets to fetch the balances for.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [x] Tests are included (if applicable)
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Preview deployment works as expected
- [ ] Ready for production